moved changes to /advisor. still getting same error
