+++ advisor/urls.py
@@ -1,0 +1,12 @@
+from django.conf.urls import patterns, url
+from django.contrib import admin
+
+import advisor.views
+
+admin.autodiscover()
+
+urlpatterns = patterns('',
+        # viewer management
+       url(r'^addViewer/$',
+                               advisor.views.AdvisorView.as_view()),
+       )
