+++ semesterly/settings.py
@@ -164,6 +164,7 @@ INSTALLED_APPS = (
     'django.contrib.staticfiles',
     'social.apps.django_app.default',
     'django_extensions',
+    'advisor',
     'authpipe',
     'timetable',
     'exams',
